Separated main screen into a file
